move sftp client example to examples directory
