+++ examples/include.am
@@ -6,3 +6,4 @@ include examples/client/include.am
 include examples/server/include.am
 include examples/echoserver/include.am
 include examples/portfwd/include.am
+include examples/sftpclient/include.am
