+++ scripts/external.test
@@ -38,10 +38,10 @@ echo -e "Success\n"
 
 
 # not having the sftp client built in is not a failure case
-./wolfsftp/client/wolfsftp -h
+./examples/sftpclient/wolfsftp -h
 if [ $? -eq 0 ]; then
     echo "Testing wolfsftp connection to $host : "
-    echo "exit" | ./wolfsftp/client/wolfsftp -u $user -P $password -h $host -p 22
+    echo "exit" | ./examples/sftpclient/wolfsftp -u $user -P $password -h $host -p 22
     RESULT=$?
     if [ $RESULT -ne 0 ]; then
         echo -e "failed to connect\n"
