--- wolfsftp/client/include.am
@@ -1,13 +1,0 @@
-# vim:ft=automake
-# All paths should be given relative to the root
-
-if BUILD_SFTP
-noinst_PROGRAMS += wolfsftp/client/wolfsftp
-noinst_HEADERS += wolfsftp/client/sftpclient.h
-wolfsftp_client_wolfsftp_SOURCES      = wolfsftp/client/sftpclient.c
-wolfsftp_client_wolfsftp_LDADD        = src/libwolfssh.la
-wolfsftp_client_wolfsftp_DEPENDENCIES = src/libwolfssh.la
-endif
-
-dist_example_DATA+= wolfsftp/client/sftpclient.c
-DISTCLEANFILES+= wolfsftp/client/.libs/wolfsftp
