--- wolfsftp/include.am
@@ -1,5 +1,0 @@
-# vim:ft=automake
-# included from Top Level Makefile.am
-# All paths should be given relative to the root
-
-include wolfsftp/client/include.am
